+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New features:
   to prevent blocks being closed too soon.
 - Removed the Gemini 2.0 Pro model, but added Gemini 2.5 Flash preview.  Updated models to the latest versions.
 - The markdown syntax highlighter has been extended to cover tables and inline attributes.
+- The wiki pages now provide metadata about file and directory size, modification timestamps, and file permissions.
 
 Bug fixes: